Include data binding
This CL changes the init script to compile data binding ToT.
It also adds it to source directories.

We may consider changing this. The main reason for this
choice (instead of pushing data binding to prebuilts) was
to avoid pushing data binding library into that folder.
The library (the part that ships w/ the app) is distributed
via SDK manager which goes into another prebuilts folder that
is NOT included by android gradle plugin. We can also
consider changing android gradle plugin to include that
directory

Change-Id: I4da2192f80553c1b97f57c97b811a2ab21047fb9